backend/foodgram/api/filters.py
import django_filters
from rest_framework.filters import SearchFilter

from recipes.models import Ingredient, Recipe, Tag

class IngredientFilter(SearchFilter):
    search_param = 'name'

    class Meta:
        model = Ingredient
        fields = ('name')


class RecipeFilter(django_filters.FilterSet):
    tags = django_filters.ModelMultipleChoiceFilter(
        field_name='tags__slug',
        to_field_name='slug',
        queryset=Tag.objects.all()
    )
    is_favorited = django_filters.BooleanFilter(
        field_name='favorite__user',
        lookup_expr='isnull',
        exclude=True
    )
    is_in_shopping_cart = django_filters.BooleanFilter(
        field_name='shopping_cart__user',
        lookup_expr='isnull',
        exclude=True
    )

    class Meta:
        model = Recipe
        fields = ('author', 'tags', 'is_favorited', 'is_in_shopping_cart')
    
    def filter_is_favorited(self, queryset, name, value):
        if value:
            return queryset.filter(favorite__user=self.request.user)
        return queryset

    def filter_is_in_shopping_cart(self, queryset, name, value):
        if value:
            return queryset.filter(shopping_cart__user=self.request.user)
        return queryset
